p4i integration Integrate 216763 from 4.1-optimized to main: Integrate 216762 from 4.1 to 4.1-optimized: Fixes:    Enable 'from' argument to be INT_MAX and still work as expected without a overflow. RevBy:    jasmin Details:
[git-p4: depot-paths = "//depot/qt/main/": change = 216764]

+++ src/corelib/tools/qstring.cpp
@@ -1923,7 +1923,7 @@ int QString::indexOf(const QString &str, int from, Qt::CaseSensitivity cs) const
     const int sl = str.d->size;
     if (from < 0)
         from += l;
-    if (sl + from > l)
+    if (uint(sl + from) > (uint)l)
         return -1;
     if (!sl)
         return from;
